requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-huggingface
huggingface_hub
transformers
torch
supabase
python-dotenv